cleanup: reset repo history
Check out the `early-projects` repo for the full, unabridged commit
history of my personal website. It spans back many years, back to when I
was first learning how to program, and before I really understood how to
use Git properly. As such, my personal website shared a repo with a
bunch of early projects that are of... er... variable quality.

To clean things up a bit, I'm hitting "reset" on this repo, and pushing
all those early projects into a new repo, keeping this repo small and
focused, with just the code that runs my main page.