cleanup and tweak CSS, with big changes to mobile
no more hacky :after element to see text on mobile. It just shows the
text box.

Also, I haven't touched this SASS in a loooong time, and it was not the
easiest to modify... I added in some variables to make life easier
later, but damn, past me really didn't help future me.